requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
PyPDF2>=3.0.0
scikit-learn>=1.4.0
pandas>=2.0.0
plotly>=5.18.0
python-docx>=1.1.0
python-dotenv>=1.0.1

